moved some classses into an Entities folder in the webapi project
