+++ src/dotnet/DocApi/Entities/MessageRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace WebApi.Entities
+{
+    public record MessageRequest
+    {
+        public required string Message { get; set; }
+    }
+}
