+++ src/dotnet/DocApi/Entities/ResponseContext.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace WebApi.Helpers
+namespace WebApi.Entities
 {
     public record ResponseContext(
        [property: JsonPropertyName("followup_questions")] string[] FollowupQuestions)
